% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_cellindex.R
\name{gts_cellindex}
\alias{gts_cellindex}
\title{List CellIndices in the GridTimeSeries data for a polygon}
\usage{
gts_cellindex(geometry)
}
\arguments{
\item{geometry}{A sf polygon or sf points. Should be in coordinate system EPSG:25833.}
}
\value{
A vector with the CellIndices for the grid cells in the GridTimeSeries data which intersects with the geometry.
}
\description{
Looks up the CellIndex for each grid cell in the GridTimeSeries data which intersects with a polygon or a set of points.
}
\examples{
## Create sf points and a sf polygon
library(sf)
xx <- c(58200, 74225, 58200, 74226)
yy <- c(6708225, 6708225, 6715025, 6715025)
points <- st_as_sf(x = data.frame(x = xx, y = yy), coords = c("x", "y"), crs = 25833)
pol <- st_as_sf(st_as_sfc(st_bbox(points)))
#' ## Extract cellindices for points
gts_cellindex(geometry = points)
## Extract cellindices for polygon
gts_cellindex(geometry = pol)
}
