% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_create_grid.R
\name{gts_create_grid}
\alias{gts_create_grid}
\title{Create the 1x1 km raster in the GridTimeSeries data}
\usage{
gts_create_grid()
}
\value{
An empty stars raster in resolution 1x1 km, identical to the one used in the GridTimeSeries data. Coordinate system EPSG:25833.
}
\description{
The function returns an empty stars raster with resolution 1x1 km and coordinate system EPSG:25833. This is identical to the grid used in the GridTimeSeries data.
}
\examples{
## Create empty raster grid
grid <- gts_create_grid()
grid
}
