% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_dl_cellindex.R
\name{gts_dl_cellindex}
\alias{gts_dl_cellindex}
\title{Download cellwise values for CellIndex}
\usage{
gts_dl_cellindex(cellindex, env_layer, start_date, end_date, verbose = FALSE)
}
\arguments{
\item{cellindex}{A vector with cellindices for which GTS data should be downloaded.}

\item{env_layer}{The quoted abbreviation for the environmental layer to download. E.g. Daily precipitaion = "rr", Temperature =  "tm", Snow depth =  "sd".}

\item{start_date}{The start date given as: 'YYYY-MM-DD'. If querying three or one hour data, the hour should be given like this: format: 'YYYY-MM-DDT06'}

\item{end_date}{The end date. See start_date.}

\item{verbose}{Set to TRUE to print the query to the console (default FALSE).}
}
\value{
A data frame with the GTS data associated with each grid cell for each step in the time resolution, e.g. daily values or values for each hour.
}
\description{
Downloads data for each grid cell identified by their CellIndex (the index for grid cells in GridTimeSeries (GTS) data fra NVE). For the same area, this is faster than \code{\link[=gts_dl_polygon]{gts_dl_polygon()}}, as the latter first looks up the cellindices for the polygon. Time resolution is given by the choice of environmental variable (env_layer) and start_date/end_date. E.g. "tm3h" for "2023-12-01T06" is three hour data for temperature collected at 06:00 and "tm" for "2023-12-01" is the daily average temperature.
}
\examples{
## Create a sf polygon and convert to json
library(sf)
xx <- c(58200, 74225, 58200, 74226)
yy <- c(6708225, 6708225, 6715025, 6715025)
pol <- st_as_sf(x = data.frame(x = xx, y = yy), coords = c("x", "y"), crs = 25833)
pol <- st_as_sf(st_as_sfc(st_bbox(pol)))
## Look up cellindices
cells <- gts_cellindex(geometry = pol)
## Download data
gts_dl_cellindex(cellindex = cells, env_layer = "tm", start_date = "2023-12-01", end_date = "2023-12-01")
gts_dl_cellindex(cellindex = cells, env_layer = "tm3h", start_date = "2023-12-01T06", end_date = "2023-12-01T09")
}
\seealso{
\code{\link[=gts_cellindex]{gts_cellindex()}} to find the cellindices for all raster cells in a polygon.
}
