% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_dl_cellindex_aggr.R
\name{gts_dl_cellindex_aggr}
\alias{gts_dl_cellindex_aggr}
\title{Download aggregated values for CellIndex}
\usage{
gts_dl_cellindex_aggr(
  cellindex,
  env_layer,
  start_date,
  end_date,
  method,
  verbose = FALSE
)
}
\arguments{
\item{cellindex}{A vector with cellindices for which GTS data should be downloaded.}

\item{env_layer}{The quoted abbreviation for the environmental layer to download. E.g. Daily precipitaion = "rr", Temperature =  "tm", Snow depth =  "sd".}

\item{start_date}{The start date given as: 'YYYY-MM-DD'. If querying three or one hour data, the hour should be given like this: format: 'YYYY-MM-DDT06'}

\item{end_date}{The end date. See start_date.}

\item{method}{The method for aggregating values ("sum", "min", "max", "avg" or "median")}

\item{verbose}{Set to TRUE to print the query to the console (default FALSE).}
}
\value{
description
}
\description{
Downloads and returns data aggregated for the chosen time resolution over grid cells identified by their CellIndex. For the same area, this is faster than \code{\link[=gts_dl_polygon_aggr]{gts_dl_polygon_aggr()}}, as the latter first looks up the cellindices for the polygon. Available methods for aggregating are sum, min, max, avg and median. Time resolution is given by the choice of environmental variable (env_layer) and start_date/end_date. E.g. "tm3h" for "2023-12-01T06" is three hour data for temperature collected at 06:00 and "tm" for "2023-12-01" is the daily average temperature. \emph{Warning: This function is very slow for 1 and 3 hour time steps as the API requires one request per time step when not requesting daily time steps.}
}
\examples{
# Example
## Create a sf polygon and convert to json
library(sf)
xx <- c(58200, 74225, 58200, 74226)
yy <- c(6708225, 6708225, 6715025, 6715025)
pol <- st_as_sf(x = data.frame(x = xx, y = yy), coords = c("x", "y"), crs = 25833)
pol <- st_as_sf(st_as_sfc(st_bbox(pol)))
pol <- sf_to_json(pol)
## Look up cellindices
cells <- gts_cellindex(polygon = pol)
## Download data
gts_dl_cellindex_aggr(cellindex = cells, env_layer = "tm", start_date = "2023-12-01", end_date = "2023-12-03", method = "avg")
gts_dl_cellindex_aggr(cellindex = cells, env_layer = "tm1h", start_date = "2023-12-01T06", end_date = "2023-12-01T12", method = "avg")
gts_dl_cellindex_aggr(cellindex = cells, env_layer = "tm3h", start_date = "2023-12-01T06", end_date = "2023-12-01T12", method = "avg")
}
\seealso{
\code{\link[=gts_cellindex]{gts_cellindex()}} to find the cellindices for all raster cells in a polygon.
}
