% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_parameters.R
\name{gts_parameters}
\alias{gts_parameters}
\title{List parameters available in the GridTimeSeries data}
\usage{
gts_parameters()
}
\value{
A table with details the parameters that are available in the GridTimeSeries data. Use the variable names in the table when downloading data.
}
\description{
Accesses the available parameters in the GridTimeSeries data and provides a table with the name of parameters, their time resolution, units of measurement and the years of available data.
}
\examples{
## List parameters
gts_parameters()
}
