% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgts-package.R
\docType{package}
\name{rgts-package}
\alias{rgts}
\alias{rgts-package}
\title{rgts: Conveniently download gridded time series data for Norway from NVE}
\description{
A set of functions for accessing and downloading the GridTimeSeries (GTS) data (api.nve.no/doc/gridtimeseries-data-gts/) from the Norwegian Water Resources and Energy Directorate (NVE). This data reposiory contains 1x1 km gridded time series data for Norway and includes several weather and environmental parameters, such as temperature, precipitation, wind direction and strength and snow depth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomas-kvalnes/rgts}
  \item Report bugs at \url{https://github.com/thomas-kvalnes/rgts/issues}
}

}
\author{
\strong{Maintainer}: Thomas Kvalnes \email{thomas.kvalnes@nina.no}

}
\keyword{internal}
