% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_dl_coord.R
\name{gts_dl_coords}
\alias{gts_dl_coords}
\title{Download values at given coordinates}
\usage{
gts_dl_coords(
  coords,
  parameter,
  start_date,
  end_date,
  return_raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{coords}{A data frame with x and y coordinates (coordinate system: (EPSG: 25833))}

\item{parameter}{The quoted abbreviation for the parameter to download. E.g. Daily precipitaion = "rr", Temperature =  "tm", Snow depth =  "sd".}

\item{start_date}{The start date given as: 'YYYY-MM-DD'. If querying three or one hour data, the hour should be given like this: format: 'YYYY-MM-DDT06'}

\item{end_date}{The end date. See start_date.}

\item{return_raw}{Set to TRUE to return the raw results from the query (a list) without coercing to a data.frame.}

\item{verbose}{Set to TRUE to print the query to the console (default FALSE).}
}
\value{
A data frame with the GTS data associated with each point for each step in the time resolution, e.g. daily values or values for each hour.
}
\description{
Downloads data for a point (coordinate) in a grid cell (coordinate system: (EPSG: 25833)) at the given time resolution. Several coordinates can be provided at once. Time resolution is given by the choice of parameter and start_date/end_date.  E.g. "tm3h" for "2023-12-01T06" is three hour data for temperature collected at 06:00 and "tm" for "2023-12-01" is the daily average temperature.\cr \cr
\emph{Note 1: The API has a cap at ~635'000 values in each download (lower if many missing values) and ends with an error for larger queries.} \cr \cr
\emph{Note 2: The API ends with an error if some of the points have the same cellindex. Use \code{\link[=gts_cellindex]{gts_cellindex()}} to check.}
}
\examples{
## Create a data frame with coordinates
coords <- data.frame(x = c(58200, 74225), y = c(6708225, 6715025))
## Download data
gts_dl_coords(coords = coords, parameter = "tm", start_date = "2023-12-01", end_date = "2023-12-02")
gts_dl_coords(coords = coords, parameter = "tm3h", start_date = "2023-12-01T23", end_date = "2023-12-02T06")
head(gts_dl_coords(coords = coords, parameter = "tm1h", start_date = "2023-12-01T23", end_date = "2023-12-02T06"))
}
